use declare_resource instead of the provider property
provider is bad because its a leaky abstraction and you don't wind up
with what you expect.  with declare_resource we will wind up with a
Chef::Resource::RpmPackage wrapping a Chef::Resource::Package::Rpm.  by
forcing the provider on Chef-11 you would get a Chef::Resource::Package
wrapping a Chef::Resource::Package::Rpm, while on Chef-12 you now get
a Chef::Resource::YumPackage wrapping a Chef::Resource::Package::Rpm.

the latter case has gotten worse in Chef-12, but that is not a bug, the
correct way to fix it is to use declare_resource.